+++ builders/MaintenanceBuilder.d.ts
@@ -1,0 +1,6 @@
+export default class MaintenanceBuilder {
+    protected config: Record<string, any>;
+    constructor();
+    isMaintenanceMode(): Promise<boolean>;
+    getData(): Promise<any>;
+}