+++ builders/MaintenanceBuilder.js
@@ -1,0 +1,13 @@
+import AppConfig from "../config/app";
+export default class MaintenanceBuilder {
+    config;
+    constructor() {
+        this.config = AppConfig.maintenance;
+    }
+    async isMaintenanceMode() {
+        return Bun.file(this.config.file).exists();
+    }
+    async getData() {
+        return Bun.file(this.config.file).json();
+    }
+}