+++ config/app.js
@@ -1,7 +1,7 @@
-import App from "../facades/App";
+import PathBuilder from "../builders/PathBuilder";
 const config = {
     maintenance: {
-        file: App.storagePath("framework/maintenance.down.json"),
+        file: new PathBuilder().storagePath("framework/maintenance.down.json"),
     }
 };
 export default config;