+++ facades/App.d.ts
@@ -1,19 +1,6 @@
+import MaintenanceBuilder from "../builders/MaintenanceBuilder";
+import PathBuilder from "../builders/PathBuilder";
 export default class App {
-    /**
-     * Path Manager
-     */
-    static appPath(path?: string): string;
-    static basePath(path?: string): string;
-    static commandsPath(path?: string): string;
-    static configPath(path?: string): string;
-    static databasePath(path?: string): string;
-    static publicPath(path?: string): string;
-    static resourcesPath(path?: string): string;
-    static routesPath(path?: string): string;
-    static rootPath(path?: string): string;
-    static storagePath(path?: string): string;
-    /**
-     * Utils
-     */
-    static isMaintenanceMode(): Promise<boolean>;
+    static get Path(): PathBuilder;
+    static get Maintenance(): MaintenanceBuilder;
 }