+++ src/config/app.ts
@@ -1,8 +1,8 @@
-import App from "@/facades/App";
+import PathBuilder from "@/builders/PathBuilder";
 
 const config: Record<string, any> = {
     maintenance: {
-        file: App.storagePath("framework/maintenance.down.json"),
+        file: new PathBuilder().storagePath("framework/maintenance.down.json"),
     }
 };