+++ src/facades/App.ts
@@ -1,53 +1,12 @@
-import AppBuilder from "@/builders/AppBuilder";
+import MaintenanceBuilder from "@/builders/MaintenanceBuilder";
+import PathBuilder from "@/builders/PathBuilder";
 
 export default class App {
-    /**
-     * Path Manager
-     */
-    public static appPath(path?: string): string {
-        return new AppBuilder().appPath(path);
+    public static get Path(): PathBuilder {
+        return new PathBuilder();
     }
 
-    public static basePath(path?: string): string {
-        return new AppBuilder().basePath(path);
-    }
-
-    public static commandsPath(path?: string): string {
-        return new AppBuilder().commandsPath(path);
-    }
-
-    public static configPath(path?: string): string {
-        return new AppBuilder().configPath(path);
-    }
-
-    public static databasePath(path?: string): string {
-        return new AppBuilder().databasePath(path);
-    }
-
-    public static publicPath(path?: string): string {
-        return new AppBuilder().publicPath(path);
-    }
-
-    public static resourcesPath(path?: string): string {
-        return new AppBuilder().resourcesPath(path);
-    }
-
-    public static routesPath(path?: string): string {
-        return new AppBuilder().routesPath(path);
-    }
-
-    public static rootPath(path?: string): string {
-        return new AppBuilder().rootPath(path);
-    }
-
-    public static storagePath(path?: string): string {
-        return new AppBuilder().storagePath(path);
-    }
-
-    /**
-     * Utils
-     */
-    public static async isMaintenanceMode(): Promise<boolean> {
-        return await new AppBuilder().isMaintenanceMode();
+    public static get Maintenance(): MaintenanceBuilder {
+        return new MaintenanceBuilder();
     }
 }